third_party/WebKit/LayoutTests/fast/events/pointerevents/mouse-pointer-capture-expected.txt

Verifies that pointer capture works for mouse.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


 ======= Set pointer capture and release implicitly  =======
    **** Move to green box & mouse press & jiggle ***** 
green received pointerover
grey received pointerenter
green received pointerenter
green received mouseover
grey received mouseenter
green received mouseenter
green received pointermove
green received mousemove
green received pointerdown
    **** Set Pointer Capture  ***** 
green received mousedown
green received gotpointercapture
green received pointermove
Capture pointer event attributes are correct!
green received mousemove
    **** Move to grey box & jiggle ***** 
green received pointermove
green received mousemove
green received pointermove
green received mousemove
    **** Move to blue box & jiggle ****
green received pointermove
green received mousemove
green received pointermove
green received mousemove
green received pointermove
green received mousemove
    **** Move back to green & again to blue & mouse release ***** 
green received pointermove
green received mousemove
green received pointermove
green received mousemove
green received pointerup
green received mouseup
green received lostpointercapture
Capture pointer event attributes are correct!
green received pointerout
green received pointerleave
grey received pointerleave
blue received pointerover
blue received pointerenter
green received mouseout
green received mouseleave
grey received mouseleave
blue received mouseover
blue received mouseenter
    **** Jiggle in blue box ***** 
blue received pointermove
blue received mousemove
    **** Move to (0,0) ***** 
blue received pointerout
blue received pointerleave
blue received mouseout
blue received mouseleave

 ======= Set pointer capture and release implicitly outside of the page =======
    **** Move to green box & mouse press & jiggle ***** 
green received pointerover
grey received pointerenter
green received pointerenter
green received mouseover
grey received mouseenter
green received mouseenter
green received pointermove
green received mousemove
green received pointerdown
    **** Set Pointer Capture  ***** 
green received mousedown
green received gotpointercapture
green received pointermove
Capture pointer event attributes are correct!
green received mousemove
    **** Move to grey box & jiggle ***** 
green received pointermove
green received mousemove
green received pointermove
green received mousemove
    **** Move to blue box & jiggle ****
green received pointermove
green received mousemove
    **** Release the mouse outside of the page & move back to blue box ****
green received lostpointercapture
green received pointerout
Capture pointer event attributes are correct!
green received pointerleave
grey received pointerleave
blue received pointerover
blue received pointerenter
green received mouseout
green received mouseleave
grey received mouseleave
blue received mouseover
blue received mouseenter
blue received pointermove
blue received mousemove
blue received pointermove
blue received mousemove
    **** Move to (0,0) ***** 
blue received pointerout
blue received pointerleave
blue received mouseout
blue received mouseleave

 ======= Set pointer capture and release explicitly  =======
    **** Move to green box & mouse press & jiggle ***** 
green received pointerover
grey received pointerenter
green received pointerenter
green received mouseover
grey received mouseenter
green received mouseenter
green received pointermove
green received mousemove
green received pointerdown
    **** Set Pointer Capture  ***** 
green received mousedown
green received gotpointercapture
green received pointermove
Capture pointer event attributes are correct!
green received mousemove
    **** Move to grey box & jiggle ***** 
green received pointermove
green received mousemove
green received pointermove
green received mousemove
    **** Move to blue box & jiggle ****
green received pointermove
    **** Release Pointer Capture  ***** 
green received mousemove
green received lostpointercapture
green received pointerout
Capture pointer event attributes are correct!
green received pointerleave
grey received pointerleave
blue received pointerover
blue received pointerenter
green received mouseout
green received mouseleave
grey received mouseleave
blue received mouseover
blue received mouseenter
blue received pointermove
blue received mousemove
blue received pointermove
blue received mousemove
    **** Move back to green & again to blue & mouse release ***** 
blue received pointerout
blue received pointerleave
green received pointerover
grey received pointerenter
green received pointerenter
blue received mouseout
blue received mouseleave
green received mouseover
grey received mouseenter
green received mouseenter
green received pointermove
green received mousemove
green received pointerout
green received pointerleave
grey received pointerleave
blue received pointerover
blue received pointerenter
green received mouseout
green received mouseleave
grey received mouseleave
blue received mouseover
blue received mouseenter
blue received pointermove
blue received mousemove
blue received pointerup
blue received mouseup
    **** Jiggle in blue box ***** 
blue received pointermove
blue received mousemove
    **** Move to (0,0) ***** 
blue received pointerout
blue received pointerleave
blue received mouseout
blue received mouseleave

 ======= Set pointer capture and remove element  =======
    **** Move to green box & mouse press & jiggle ***** 
green received pointerover
grey received pointerenter
green received pointerenter
green received mouseover
grey received mouseenter
green received mouseenter
green received pointermove
green received mousemove
green received pointerdown
    **** Set Pointer Capture  ***** 
green received mousedown
green received gotpointercapture
green received pointermove
Capture pointer event attributes are correct!
green received mousemove
    **** Move to grey box & jiggle ***** 
green received pointermove
green received mousemove
green received pointermove
green received mousemove
    **** Move to blue box & jiggle ****
green received pointermove
    **** Remove Element ***** 
grey received mousemove
document received lostpointercapture
blue received pointerover
blue received pointerenter
blue received mouseover
blue received mouseenter
blue received pointermove
blue received mousemove
blue received pointermove
blue received mousemove
    **** Move back to green & again to blue & mouse release ***** 
blue received pointerout
blue received pointerleave
grey received pointerover
grey received pointerenter
blue received mouseout
blue received mouseleave
grey received mouseover
grey received mouseenter
grey received pointermove
grey received mousemove
grey received pointerout
grey received pointerleave
blue received pointerover
blue received pointerenter
grey received mouseout
grey received mouseleave
blue received mouseover
blue received mouseenter
blue received pointermove
blue received mousemove
blue received pointerup
blue received mouseup
    **** Jiggle in blue box ***** 
blue received pointermove
blue received mousemove
    **** Move to (0,0) ***** 
blue received pointerout
blue received pointerleave
blue received mouseout
blue received mouseleave
PASS successfullyParsed is true

TEST COMPLETE
